refactor(runtime): isolate systemd reconciliation
