+++ packages/cli/src/runtime/converge-lock.test.ts
@@ -2,7 +2,7 @@ import { afterEach, describe, expect, test } from "bun:test";
 import { mkdtempSync, rmSync } from "node:fs";
 import { tmpdir } from "node:os";
 import { join } from "node:path";
-import { withRuntimeConvergeLockAsync } from "./manifest";
+import { withRuntimeConvergeLockAsync } from "./converge-lock";
 import { getRuntimePaths } from "./paths";
 
 const originalEnv = { ...process.env };
